handle errors returned from Limiter
The limiter will return an error if the requested rate is larger
than the burst limit or if the context has a deadline that is
earlier then when the requested rate would be allowed.  For the
first case, we return a EntityTooLarge http error (413). For the
second case we return the StatusTooManyRequests error (429).